requirements.txt
certifi==2017.11.5
chardet==3.0.4
gitdb2==2.0.3
GitPython==2.1.8
idna==2.6
requests==2.18.4
smmap2==2.0.3
tqdm==4.19.5
urllib3==1.22
